requirements.txt
azure-identity
azure-appconfiguration
azure-keyvault-secrets
